+++ ogma-core/templates/ros/copilot/src/copilot_logger.cpp
@@ -1,0 +1,27 @@
+#include <functional>
+#include <memory>
+
+#include "rclcpp/rclcpp.hpp"
+
+#include "std_msgs/msg/empty.hpp"
+using std::placeholders::_1;
+
+class CopilotLogger : public rclcpp::Node {
+  public:
+    CopilotLogger() : Node("copilotlogger") {
+{{{logMsgSubscriptionS}}}
+    }
+
+  private:
+{{{logMsgCallbacks}}}
+{{{logMsgSubscriptionDeclrs}}}
+};
+
+int main(int argc, char* argv[]) {
+  rclcpp::init(argc, argv);
+  rclcpp::spin(std::make_shared<CopilotLogger>());
+  rclcpp::shutdown();
+  return 0;
+}
+
+