+++ ogma-extra/CHANGELOG.md
@@ -1,5 +1,10 @@
 # Revision history for ogma-extra
 
+## [1.5.0] - 2024-11-21
+
+* Version bump 1.5.0 (#178).
+* Introduce template expansion functionality (#162).
+
 ## [1.4.1] - 2024-09-21
 
 * Version bump 1.4.1 (#155).