+++ ogma-language-jsonspec/CHANGELOG.md
@@ -1,5 +1,10 @@
 # Revision history for ogma-language-jsonspec
 
+## [1.5.0] - 2024-11-21
+
+* Version bump 1.5.0 (#178).
+* Make JSONFormat readable (#170).
+
 ## [1.4.1] - 2024-09-21
 
 * Version bump 1.4.1 (#155).